moved form into Postcode Form Component
